requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
redis==5.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0

# AI/ML Libraries
openai==1.3.7
transformers==4.30.2
torch==2.1.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# AWS Integration
boto3==1.34.0
botocore==1.34.0

# Message Queue
pika==1.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
click==8.1.7 
huggingface_hub==0.16.4 